implement create event based fixture and usage in test

+++ tests/api/Booking/bookEvent.spec.ts
@@ -1,50 +1,10 @@
-import { expect, test } from "../../../fixtures/ApiFixture";
-import { cleanupEvent } from "../../../utils/CleanUpHelper";
-import { CreateEventByApi } from "../../../utils/CreateEventHelperApi";
-import { generateBookingPayload } from "../../../utils/factories/bookingFactory";
+    import { expect, test } from "../../../fixtures/ApiFixture";
+    import { generateBookingPayload } from "../../../utils/factories/bookingFactory";
 
+    test('@api-booking should be able to book default event successfully', async ({ bookingService }) => {
 
-test('@api-booking should be able to book default event successfully', async ({ bookingService }) => {
-
-    // using default event as to create booking 
-    const bookingPayload = generateBookingPayload(1, 1);
-
-    test.info().attach("Create booking payload",
-        {
-            body: JSON.stringify(bookingPayload, null, 2),
-            contentType: "application/json"
-        }
-    );
-
-    const response = await bookingService.createBooking(bookingPayload);
-
-    test.info().attach("Create booking Response",
-        {
-            body: JSON.stringify(response.body, null, 2),
-            contentType: "application/json"
-        }
-    );
-
-    expect(response.status).toBe(201);
-
-    if (response.status === 201) {
-        expect(response.body.success).toBeTruthy();
-        expect(response.body.message).toBe("Booking confirmed!");
-    }
-})
-
-test("@api-booking should be able to book newly created event ", async ({ eventService, bookingService }) => {
-    let eventId: number | undefined;
-    try {
-        const { eventResponse } = await CreateEventByApi(eventService);
-
-        if (eventResponse.status !== 201) {
-            throw new Error("Unable to create event")
-        }
-
-        eventId = eventResponse.body.data.id;
-        const seatsCount = eventResponse.body.data.totalSeats;
-        const bookingPayload = generateBookingPayload(eventId, 1);
+        // using default event as to create booking 
+        const bookingPayload = generateBookingPayload(3, 1);
 
         test.info().attach("Create booking payload",
             {
@@ -53,14 +13,51 @@ test("@api-booking should be able to book newly created event ", async ({ eventS
             }
         );
 
-        const bookingResponse = await bookingService.createBooking(bookingPayload);
+        const response = await bookingService.createBooking(bookingPayload);
 
         test.info().attach("Create booking Response",
+            {
+                body: JSON.stringify(response.body, null, 2),
+                contentType: "application/json"
+            }
+        );
+
+        expect(response.status).toBe(201);
+
+        if (response.status === 201) {
+            expect(response.body.success).toBeTruthy();
+            expect(response.body.message).toBe("Booking confirmed!");
+        }
+    })
+
+    test("@api-booking should be able to book newly created event ", async ({ eventResource, eventService, bookingService }) => {
+
+        const eventId = eventResource.id;
+        const seatsCount = eventResource.totalSeats;
+
+        const bookingPayload = generateBookingPayload(eventId, 1);
+
+        await test.info().attach("Create booking payload",
+            {
+                body: JSON.stringify(bookingPayload, null, 2),
+                contentType: "application/json"
+            }
+        );
+
+        const bookingResponse = await bookingService.createBooking(bookingPayload);
+
+        await test.info().attach("Create booking Response",
             {
                 body: JSON.stringify(bookingResponse.body, null, 2),
                 contentType: "application/json"
             }
         );
+
+        expect(bookingResponse.status).toBe(201);
+        if (bookingResponse.status === 201) {
+            expect(bookingResponse.body.success).toBeTruthy();
+            expect(bookingResponse.body.message).toBe("Booking confirmed!");
+        }
 
         const getResponse = await eventService.getEvent(eventId);
         if (getResponse.status !== 200) {
@@ -69,33 +66,16 @@ test("@api-booking should be able to book newly created event ", async ({ eventS
 
         const updatedCount = getResponse.body.data.availableSeats;
 
-        expect(bookingResponse.status).toBe(201);
-        if (bookingResponse.status === 201) {
-            expect(bookingResponse.body.success).toBeTruthy();
-            expect(bookingResponse.body.message).toBe("Booking confirmed!");
-        }
-
         expect(updatedCount).toBe(seatsCount - 1);
-    } finally {
-        await cleanupEvent(eventService, eventId);
-    }
-});
+    });
 
 
-test("@api-booking should be able to retrieve booking by reference id", async ({ eventService, bookingService }) => {
+    test("@api-booking should be able to retrieve booking by reference id", async ({ eventResource, bookingService }) => {
 
-    let eventId: number | undefined;
-    try {
-        const { eventResponse } = await CreateEventByApi(eventService);
-
-        if (eventResponse.status !== 201) {
-            throw new Error("Unable to create event")
-        }
-
-        eventId = eventResponse.body.data.id;
+        const eventId = eventResource.id;
 
         const bookingPayload = generateBookingPayload(eventId, 1);
-        test.info().attach("Create booking payload",
+        await test.info().attach("Create booking payload",
             {
                 body: JSON.stringify(bookingPayload, null, 2),
                 contentType: "application/json"
@@ -103,7 +83,7 @@ test("@api-booking should be able to retrieve booking by reference id", async ({
         );
 
         const bookingResponse = await bookingService.createBooking(bookingPayload);
-        test.info().attach("Create booking Response",
+        await test.info().attach("Create booking Response",
             {
                 body: JSON.stringify(bookingResponse.body, null, 2),
                 contentType: "application/json"
@@ -119,26 +99,15 @@ test("@api-booking should be able to retrieve booking by reference id", async ({
         const getResponse = await bookingService.getBookingByRefId(referenceId);
         expect(getResponse.status).toBe(200);
         expect(getResponse.body.success).toBeTruthy();
-
-    } finally {
-        await cleanupEvent(eventService, eventId);
-    }
-});
+    });
 
 
-test("@api-booking create booking with zero tickets count", async ({ eventService, bookingService }) => {
-    let eventId: number | undefined;
-    try {
-        const { eventResponse } = await CreateEventByApi(eventService);
+    test("@api-booking create booking with zero tickets count", async ({ eventResource, bookingService }) => {
 
-        if (eventResponse.status !== 201) {
-            throw new Error("Unable to create event")
-        }
-
-        eventId = eventResponse.body.data.id;
+        const eventId = eventResource.id;
         const bookingPayload = generateBookingPayload(eventId, 0);
 
-        test.info().attach("Create booking payload",
+        await test.info().attach("Create booking payload",
             {
                 body: JSON.stringify(bookingPayload, null, 2),
                 contentType: "application/json"
@@ -147,38 +116,27 @@ test("@api-booking create booking with zero tickets count", async ({ eventServic
 
         const bookingResponse = await bookingService.createBooking(bookingPayload);
 
-        test.info().attach("Create booking Response",
+        await test.info().attach("Create booking Response",
             {
                 body: JSON.stringify(bookingResponse.body, null, 2),
                 contentType: "application/json"
             }
         );
 
+        expect(bookingResponse.status).toBe(400);
         if (bookingResponse.status === 400) {
             expect(bookingResponse.body.success).toBeFalsy();
             expect(bookingResponse.body.error).toBe("Validation failed");
             expect(bookingResponse.body.details.at(0)?.message).toBe("Quantity must be an integer between 1 and 10")
         }
+    });
 
-    } finally {
-        await cleanupEvent(eventService, eventId);
-    }
-});
+    test("@api-booking create booking with max count +1", async ({ bookingService, eventResource }) => {
 
-test("@api-booking create booking with max count +1", async ({ bookingService, eventService }) => {
-
-    let eventId: number | undefined;
-    try {
-        const { eventResponse } = await CreateEventByApi(eventService);
-
-        if (eventResponse.status !== 201) {
-            throw new Error("Unable to create event")
-        }
-
-        eventId = eventResponse.body.data.id;
+        const eventId = eventResource.id;
         const bookingPayload = generateBookingPayload(eventId, 11);
 
-        test.info().attach("Create booking payload",
+        await test.info().attach("Create booking payload",
             {
                 body: JSON.stringify(bookingPayload, null, 2),
                 contentType: "application/json"
@@ -187,7 +145,7 @@ test("@api-booking create booking with max count +1", async ({ bookingService, e
 
         const bookingResponse = await bookingService.createBooking(bookingPayload);
 
-        test.info().attach("Create booking Response",
+        await test.info().attach("Create booking Response",
             {
                 body: JSON.stringify(bookingResponse.body, null, 2),
                 contentType: "application/json"
@@ -201,29 +159,17 @@ test("@api-booking create booking with max count +1", async ({ bookingService, e
             expect(bookingResponse.body.error).toBe("Validation failed");
             expect(bookingResponse.body.details.at(0)?.message).toBe("Quantity must be an integer between 1 and 10")
         }
-
-    } finally {
-        await cleanupEvent(eventService, eventId);
-    }
-});
+    });
 
 
-test("@api-booking should be able to book event with max ticket count 10", async ({ eventService, bookingService }) => {
+    test("@api-booking should be able to book event with max ticket count 10", async ({ eventResource, eventService, bookingService }) => {
 
-    let eventId: number | undefined;
-    try {
-        const { eventResponse } = await CreateEventByApi(eventService);
-
-        if (eventResponse.status !== 201) {
-            throw new Error("Unable to create event")
-        }
-
-        eventId = eventResponse.body.data.id;
-        const seatsCount = eventResponse.body.data.totalSeats;
-        const seatPerPrice = eventResponse.body.data.price;
+        const eventId = eventResource.id;
+        const seatsCount = eventResource.totalSeats;
+        const seatPerPrice = eventResource.price;
         const bookingPayload = generateBookingPayload(eventId, 10);
 
-        test.info().attach("Create booking payload",
+        await test.info().attach("Create booking payload",
             {
                 body: JSON.stringify(bookingPayload, null, 2),
                 contentType: "application/json"
@@ -232,20 +178,12 @@ test("@api-booking should be able to book event with max ticket count 10", async
 
         const bookingResponse = await bookingService.createBooking(bookingPayload);
 
-        test.info().attach("Create booking Response",
+        await test.info().attach("Create booking Response",
             {
                 body: JSON.stringify(bookingResponse.body, null, 2),
                 contentType: "application/json"
             }
         );
-
-
-        const getResponse = await eventService.getEvent(eventId);
-        if (getResponse.status !== 200) {
-            throw new Error("Failed get update seat count from get API");
-        }
-
-        const updatedCount = getResponse.body.data.availableSeats;
 
         expect(bookingResponse.status).toBe(201);
         if (bookingResponse.status === 201) {
@@ -253,31 +191,22 @@ test("@api-booking should be able to book event with max ticket count 10", async
             expect(bookingResponse.body.message).toBe("Booking confirmed!");
             expect(Number(bookingResponse.body.data.totalPrice)).toBe(seatPerPrice * 10);
         }
-
-        expect(updatedCount).toBe(seatsCount - 10);
-
-    } finally {
-        await cleanupEvent(eventService, eventId);
-    }
-
-})
-
-test("@api-booking should be able to cancel the booking", async ({ eventService, bookingService }) => {
-
-    let eventId: number | undefined
-
-    try {
-        const { eventResponse } = await CreateEventByApi(eventService);
-
-        if (eventResponse.status !== 201) {
-            throw new Error("Unable to create event")
+        const getResponse = await eventService.getEvent(eventId);
+        if (getResponse.status !== 200) {
+            throw new Error("Failed get update seat count from get API");
         }
-        expect(eventResponse.status).toBe(201)
 
-        eventId = eventResponse.body.data.id;
+        const updatedCount = getResponse.body.data.availableSeats;
+        expect(updatedCount).toBe(seatsCount - 10);
+    })
+
+    test("@api-booking should be able to cancel the booking and restroe seats", async ({ eventService, eventResource, bookingService }) => {
+
+        const eventId = eventResource.id;
+        const seatsCount = eventResource.totalSeats;
 
         const bookingPayload = generateBookingPayload(eventId, 1);
-        test.info().attach("Create booking payload",
+        await test.info().attach("Create booking payload",
             {
                 body: JSON.stringify(bookingPayload, null, 2),
                 contentType: "application/json"
@@ -285,7 +214,7 @@ test("@api-booking should be able to cancel the booking", async ({ eventService,
         );
 
         const bookingResponse = await bookingService.createBooking(bookingPayload);
-        test.info().attach("Create booking Response",
+        await test.info().attach("Create booking Response",
             {
                 body: JSON.stringify(bookingResponse.body, null, 2),
                 contentType: "application/json"
@@ -300,8 +229,22 @@ test("@api-booking should be able to cancel the booking", async ({ eventService,
         expect(bookingResponse.body.success).toBeTruthy();
         expect(bookingResponse.body.message).toBe("Booking confirmed!");
 
+        const getResponse = await eventService.getEvent(eventId);
+        if (getResponse.status !== 200) {
+            throw new Error("Failed get update seat count from get API");
+        }
+
+        const updatedCount = getResponse.body.data.availableSeats;
+        expect(updatedCount).toBe(seatsCount - 1);
+
         const bookingId = bookingResponse.body.data.id
         const deleteResponse = await bookingService.deleteBooking(bookingId);
+        await test.info().attach("Create booking Response",
+            {
+                body: JSON.stringify(deleteResponse.body, null, 2),
+                contentType: "application/json"
+            }
+        );
 
         if (deleteResponse.status !== 200) {
             throw new Error(`Unable to cancel the ticker for event :${eventId}`);
@@ -310,8 +253,52 @@ test("@api-booking should be able to cancel the booking", async ({ eventService,
         expect(deleteResponse.status).toBe(200);
         expect(deleteResponse.body.message).toBe("Booking cancelled");
 
-    } finally {
-        await cleanupEvent(eventService, eventId);
-    }
+        const getResponse2 = await eventService.getEvent(eventId);
+        if (getResponse2.status !== 200) {
+            throw new Error("Failed get restore seat count from get API");
+        }
 
-})
+        const restoreSeatCount = getResponse2.body.data.availableSeats;
+        expect(seatsCount).toBe(restoreSeatCount)
+    })
+
+
+    test("@api-booking should not able to book event on unknown event", async ({ bookingService }) => {
+
+        const eventId = 24525;
+        const bookingPayload = generateBookingPayload(eventId, 2);
+
+        await test.info().attach("Create booking payload",
+            {
+                body: JSON.stringify(bookingPayload, null, 2),
+                contentType: "application/json"
+            }
+        );
+
+        const bookingResponse = await bookingService.createBooking(bookingPayload);
+
+        await test.info().attach("Create booking Response",
+            {
+                body: JSON.stringify(bookingResponse.body, null, 2),
+                contentType: "application/json"
+            }
+        );
+
+        expect(bookingResponse.status).toBe(404);
+
+        if (bookingResponse.status === 404) {
+            expect(bookingResponse.body.success).toBeFalsy();
+            expect(bookingResponse.body.error).toBe(`Event with id ${eventId} not found`);
+        }
+    })
+
+    test("should return 401 for invalid token", async ({unauthorizedBookingService, eventResource}) => {
+
+        const payload = generateBookingPayload(eventResource.id, 1);
+
+        const response = await unauthorizedBookingService.createBooking(payload);
+
+        expect(response.status).toBe(401);
+    }
+    );
+
